Added EEPROM_writeData

+++ SENSE_util/EEPROM.bak.Data.h
@@ -91,10 +91,10 @@ uint8_t EEPROM_readByte( uint32_t address, uint8_t ACK );
 uint8_t EEPROM_writeByte( uint32_t address, uint8_t data, uint8_t ACK );
 uint8_t EEPROM_writeData( uint32_t address, uint8_t * bpData, uint8_t length, uint8_t ACK);
 uint8_t EEPROM_readData( uint32_t address, uint8_t * bpData, uint8_t lenght, uint8_t ACK );
-uint8_t * EEPROM_readPage( uint16_t pageNumber );
-uint8_t EEPROM_writePage( uint16_t pageNumber, uint8_t * data );
-uint8_t EEPROM_sequentialRead( uint32_t address, uint16_t numOfBytes, uint8_t * dest, uint8_t ACK );
-uint8_t EEPROM_sequentialWrite( uint32_t address, uint16_t numOfBytes, uint8_t * data );
+uint8_t * EEPROM_readPage( unsigned int pageNumber );
+uint8_t EEPROM_writePage( unsigned int pageNumber, uint8_t * data );
+uint8_t EEPROM_sequentialRead( uint16_t address, uint16_t numOfBytes, uint8_t * dest, uint8_t ACK );
+uint8_t EEPROM_sequentialWrite( uint16_t address, uint16_t numOfBytes, uint8_t * data );
 uint8_t EEPROM_erase( void );
 
 #endif // EEPROM_H_
